moved byte conversion methods into their own header file
